Moved excerise_card.dart file
